+++ pjrpc/__init__.py
@@ -15,7 +15,7 @@ from pjrpc.__about__ import (
     __version__,
     __author__,
     __email__,
-    __license__
+    __license__,
 )
 
 # shortcuts